refactored nav component and css
